requirements.txt
PySide6>=6.5.0
mss>=9.0.1
keyboard>=0.13.5
requests>=2.31.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
# tesserocr устанавливается вручную через .whl файл!